Папки с классами старого проекта пересены в корневой пакет
