+++ blog-app/src/main/java/io/github/tasoula/blog_app/config/WebConfiguration.java
@@ -1,0 +1,20 @@
+package io.github.tasoula.blog_app.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.web.multipart.MultipartResolver;
+import org.springframework.web.multipart.support.StandardServletMultipartResolver;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+
+@Configuration
+@EnableWebMvc
+@ComponentScan(basePackages = "io/github/tasoula")
+@PropertySource("classpath:application.properties")
+public class WebConfiguration {
+    @Bean(name = "multipartResolver")
+    public MultipartResolver multipartResolver() {
+        return new StandardServletMultipartResolver();
+    }
+}
