+++ blog-app/src/main/java/io/github/tasoula/blog_app/dto/Comment.java
@@ -1,0 +1,19 @@
+package io.github.tasoula.blog_app.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.sql.Timestamp;
+import java.util.UUID;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class Comment {
+    private UUID id;
+    private String text;
+    private Timestamp createdAt;
+}
