+++ blog-app/src/main/java/io/github/tasoula/blog_app/repository/interfaces/ImageRepository.java
@@ -1,0 +1,8 @@
+package io.github.tasoula.blog_app.repository.interfaces;
+
+import java.util.UUID;
+
+public interface ImageRepository {
+    String getPostImage(UUID postId);
+    void updatePostImage(UUID postId, String filename);
+}
