+++ blog-app/src/main/java/io/github/tasoula/blog_app/repository/interfaces/TagRepository.java
@@ -1,0 +1,12 @@
+package io.github.tasoula.blog_app.repository.interfaces;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface  TagRepository {
+    List<String> getTagsByPostId(UUID postId);
+
+    void updatePostTags(UUID postId, List<String> tags);
+
+
+}
