--- blog-app/src/main/java/io/github/tasoula/repository/interfaces/ImageRepository.java
@@ -1,8 +1,0 @@
-package io.github.tasoula.repository.interfaces;
-
-import java.util.UUID;
-
-public interface ImageRepository {
-    String getPostImage(UUID postId);
-    void updatePostImage(UUID postId, String filename);
-}
